removed the folder from the file explorer now it's only files the frontend has file explorer, it's able to create file rename them, delete files, and save them or update their contents made sure the project slug works fine made a get request to get the file content and have it appear on the code editor

+++ devkin/src/main/java/com/example/devkin/entities/File.java
@@ -12,7 +12,7 @@ public class File {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(nullable = false)
-    private Long fileId;
+    private Integer fileId;
 
     @Column(nullable = false)
     private String fileName;
@@ -45,11 +45,11 @@ public class File {
 
     // Getters and Setters
 
-    public Long getFileId() {
+    public Integer getFileId() {
         return fileId;
     }
 
-    public void setFileId(Long fileId) {
+    public void setFileId(Integer fileId) {
         this.fileId = fileId;
     }
